+++ pkg/analyzer/lib/dart/element/type_system.dart
@@ -42,8 +42,6 @@ abstract class TypeSystem {
   /// be provided.
   DartType instantiateToBounds2({
     ClassElement? classElement,
-    @Deprecated("Use 'typeAliasElement' instead")
-        FunctionTypeAliasElement? functionTypeAliasElement,
     TypeAliasElement? typeAliasElement,
     required NullabilitySuffix nullabilitySuffix,
   });